backend/requirements.txt
# === requirements.txt ===
# FastAPI backend dependencies for LFA Legacy GO

# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# Environment & Configuration  
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Additional utilities
python-dateutil>=2.8.0
email-validator>=2.1.0

# === WEATHER INTEGRATION DEPENDENCIES ===
# HTTP client for weather API requests
aiohttp>=3.9.1

# Background task processing (optional, for advanced features)
celery>=5.3.4
redis>=5.0.1

# Additional async utilities
asyncio-mqtt>=0.16.0